J'ai fait quelque modification de passage

+++ src/fr/iutvalence/info/dut/m2107/Grid.java
@@ -41,7 +41,7 @@ public class Grid {
 	 * @param x
 	 * @param y
 	 */
-	public boolean inGrid(float x, float y) {
+	public boolean inGrid(int x, int y) {
 		// TODO - implement Grid.inGrid
 		throw new UnsupportedOperationException();
 	}
